refactor: Create a separate html helper.
